Camera mode swapping added
Camera has a set of modes that are swapped through with a press of the C key, by default the camera starts in "path" mode which follows a path set for each room. Pressing C will iterate through the different modes, and reaching the end of the path will activate object-following mode centered on the ball. Will add manual controlled camera modes next.